requirements.txt
streamlit
yfinance
python-dotenv
matplotlib
google-generativeai
reportlab
